+++ packages/ui/src/components/date-time/plain-date-time-or-date-picker/index.ts
@@ -1,0 +1,2 @@
+export { default as CoarPlainDateTimeOrDatePicker } from './CoarPlainDateTimeOrDatePicker.vue';
+export type { CoarPlainDateTimeOrDatePickerSize } from './CoarPlainDateTimeOrDatePicker.vue';