+++ packages/ui/src/components/date-time/zoned-date-time-or-date-picker/index.ts
@@ -1,0 +1,2 @@
+export { default as CoarZonedDateTimeOrDatePicker } from './CoarZonedDateTimeOrDatePicker.vue';
+export type { CoarZonedDateTimeOrDatePickerSize } from './CoarZonedDateTimeOrDatePicker.vue';