+++ packages/ui/src/index.ts
@@ -274,6 +274,10 @@ export { CoarPlainDateTimePicker } from './components/date-time/plain-date-time-
 export type { CoarPlainDateTimePickerSize } from './components/date-time/plain-date-time-picker';
 export { CoarZonedDateTimePicker } from './components/date-time/zoned-date-time-picker';
 export type { CoarZonedDateTimePickerSize } from './components/date-time/zoned-date-time-picker';
+export { CoarZonedDateTimeOrDatePicker } from './components/date-time/zoned-date-time-or-date-picker';
+export type { CoarZonedDateTimeOrDatePickerSize } from './components/date-time/zoned-date-time-or-date-picker';
+export { CoarPlainDateTimeOrDatePicker } from './components/date-time/plain-date-time-or-date-picker';
+export type { CoarPlainDateTimeOrDatePickerSize } from './components/date-time/plain-date-time-or-date-picker';
 export { CoarPlainDateView } from './components/date-time/plain-date-view';
 export type { CoarPlainDateViewProps, CoarPlainDateViewSize } from './components/date-time/plain-date-view';
 export { CoarPlainDateTimeView } from './components/date-time/plain-date-time-view';